+++ src/parser/enrichers/monster/Beholder/_module.mjs
@@ -1,0 +1,1 @@
+export { default as EyeRays } from "./EyeRays.mjs";