+++ src/parser/enrichers/monster/_module.mjs
@@ -23,3 +23,4 @@ export * as Aboleth from "./Aboleth/_module.mjs";
 export * as EldritchCannon from "./EldritchCannon/_module.mjs";
 export * as Stirge from "./Stirge/_module.mjs";
 export * as SwarmOfStirges from "./SwarmOfStirges/_module.mjs";
+export * as Beholder from "./Beholder/_module.mjs";