+++ src/parser/enrichers/spell/Haste.mjs
@@ -8,10 +8,10 @@ export default class Haste extends DDBEnricherData {
       {
         changes: [
           DDBEnricherData.ChangeHelper.signedAddChange("2", 20, "system.attributes.ac.bonus"),
+          DDBEnricherData.ChangeHelper.customChange("*2", 30, "system.attributes.movement.all"),
         ],
         midiChanges: [
           DDBEnricherData.ChangeHelper.overrideChange("1", 20, "flags.midi-qol.advantage.ability.save.dex"),
-          DDBEnricherData.ChangeHelper.customChange("*2", 30, "system.attributes.movement.all"),
         ],
       },
     ];